Move ThreadSafeHeadersTests to acceptance testing, a Core feat
